Don't clear out existing policy exclusion list
Here is what the function does:
Takes 2 policies:
- one is OnlyEnvironments and is intended for Teams environments
- one is ExceptEnvironments and is intended for all other environments (excludes Teams)

The function takes a parameter (ExceptionEnvironmentIds) which is the list of environments that should be excluded from the second policy (in addition to the Teams environments).
Here's the issue:
We wipe out the existing exclusion and starts from a new list. So, if the call to get the environment from BAP around like 1360 fails (or BAP returns incomplete data or something), then that environment gets dropped from the excluded-environments list --> policy is enforced on the environment and apps/flows break.